+++ src/argon2-browser.d.ts
@@ -1,1 +1,2 @@
 declare module "argon2-browser/lib/argon2.js";
+declare module "argon2-browser/dist/argon2-bundled.min.js";